Major change: --galileo and --config_for_timing get satellite config from conf file
See etc/gpsctl.conf and etc/gpsctl.conf.default for examples

-- galileo and --config_for_timing will configure as in the example etc/gpsctl.conf if no /etc/gpsctl.conf file exists.

If the file exists, the settings in the file will be applied.  If a setting is commented out, the default for that setting as in etc/gpsctl.conf will be applied.

etc/gpsctl.conf.default contains the settings as produced by --reset.

So it's possible to use --galileo to turn galileo off, if you wish (should really rename it to --satellites)

NMEA version is set to 41 regardless if either --galileo or --config_for_timing are invoked.